+++ src/test/java/com/javaetmoi/benchmark/mapping/mapper/BullMapperTest.java
@@ -1,6 +1,5 @@
-package com.javaetmoi.benchmark.mapping.mapper.dozer;
+package com.javaetmoi.benchmark.mapping.mapper;
 
-import com.javaetmoi.benchmark.mapping.mapper.OrderMapper;
 import com.javaetmoi.benchmark.mapping.mapper.bull.BullMapper;
 
 public class BullMapperTest extends AbstractMapperTest {
