+++ src/test/java/com/javaetmoi/benchmark/mapping/mapper/DozerMapperTest.java
@@ -1,6 +1,6 @@
-package com.javaetmoi.benchmark.mapping.mapper.dozer;
+package com.javaetmoi.benchmark.mapping.mapper;
 
-import com.javaetmoi.benchmark.mapping.mapper.OrderMapper;
+import com.javaetmoi.benchmark.mapping.mapper.dozer.DozerMapper;
 
 public class DozerMapperTest extends AbstractMapperTest {
