+++ src/test/java/com/javaetmoi/benchmark/mapping/mapper/SelmaMapperTest.java
@@ -1,6 +1,5 @@
-package com.javaetmoi.benchmark.mapping.mapper.dozer;
+package com.javaetmoi.benchmark.mapping.mapper;
 
-import com.javaetmoi.benchmark.mapping.mapper.OrderMapper;
 import com.javaetmoi.benchmark.mapping.mapper.selma.SelmaMapper;
 
 public class SelmaMapperTest extends AbstractMapperTest {
